Moved LabelledLogger to Util
